Split utility functions and ModularScale class into new modules
